WIP #18518 - Add tests and improvements for RegistryApi
This moves RegistryApi away from RestClient to Docker::Connection.
It also improves the interface of the class to provide a simpler
way to query images and their tags.

To fix issues with unsupported api versions of registries it now
tries both and returns a normalized format of search results.

To query Docker Hub there is now RegistryApi.docker_hub to get an
instance back configured for it.